fix: VaccinationServiceV2에서 LoginInfo로 Pet 정보 미확인 내용 정정

+++ src/main/java/com/kau/capstone/v2/vaccination/service/VaccinationServiceV2.java
@@ -1,15 +1,18 @@
 package com.kau.capstone.v2.vaccination.service;
 
+import com.kau.capstone.entity.member.Member;
+import com.kau.capstone.entity.member.repository.MemberRepository;
 import com.kau.capstone.entity.pet.Pet;
 import com.kau.capstone.entity.pet.repository.PetRepository;
 import com.kau.capstone.v1.auth.dto.LoginInfo;
-import com.kau.capstone.v1.pet.exception.PetNotFoundException;
+import com.kau.capstone.v2.pet.exception.PetNotFoundExceptionV2;
 import com.kau.capstone.v2.vaccination.dto.CreateVaccinationReqV2;
 import com.kau.capstone.v2.vaccination.dto.PutVaccinationReqV2;
 import com.kau.capstone.v2.vaccination.dto.VaccinationsResV2;
 import com.kau.capstone.entity.vaccination.Vaccination;
 import com.kau.capstone.entity.vaccination.repository.VaccinationRepository;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,12 +22,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class VaccinationServiceV2 {
 
     private final PetRepository petRepository;
+    private final MemberRepository memberRepository;
     private final VaccinationRepository vaccinationRepository;
 
     @Transactional
     public void createVaccinationInfo(LoginInfo loginInfo, Long petId, CreateVaccinationReqV2 request) {
-        Pet pet = petRepository.findById(petId)
-            .orElseThrow(PetNotFoundException::new);
+        Member member = memberRepository.getById(loginInfo.memberId());
+        Pet pet = petRepository.getByIdAndMemberAndDeletedAtIsNull(petId, member);
 
         Vaccination vaccination = Vaccination.of(pet, request);
         vaccinationRepository.save(vaccination);
@@ -32,8 +36,8 @@ public class VaccinationServiceV2 {
 
     @Transactional(readOnly = true)
     public VaccinationsResV2 getVaccinationInfo(LoginInfo loginInfo, Long petId) {
-        Pet pet = petRepository.findById(petId)
-                .orElseThrow(PetNotFoundException::new);
+        Member member = memberRepository.getById(loginInfo.memberId());
+        Pet pet = petRepository.getByIdAndMemberAndDeletedAtIsNull(petId, member);
 
         List<Vaccination> vaccinations = vaccinationRepository.findAllByPetOrderByVaccinatedAtDesc(pet);
         return VaccinationsResV2.of(pet, vaccinations);
@@ -41,13 +45,21 @@ public class VaccinationServiceV2 {
 
     @Transactional
     public void putVaccinationInfo(LoginInfo loginInfo, Long vaccinationId, PutVaccinationReqV2 request) {
+//        Member member = memberRepository.getById(loginInfo.memberId());
+//        Pet pet = petRepository.getByIdAndMemberAndDeletedAtIsNull(petId, member);
         Vaccination vaccination = vaccinationRepository.getById(vaccinationId);
+        if (!Objects.equals(vaccination.getPet().getMember().getId(), loginInfo.memberId()))
+            throw new PetNotFoundExceptionV2();
         vaccination.modify(request);
     }
 
     @Transactional
     public void deleteVaccinationInfo(LoginInfo loginInfo, Long vaccinationId) {
+//        Member member = memberRepository.getById(loginInfo.memberId());
+//        Pet pet = petRepository.getByIdAndMemberAndDeletedAtIsNull(petId, member);
         Vaccination vaccination = vaccinationRepository.getById(vaccinationId);
+        if (!Objects.equals(vaccination.getPet().getMember().getId(), loginInfo.memberId()))
+            throw new PetNotFoundExceptionV2();
         vaccinationRepository.delete(vaccination);
     }
